requirements-dev.txt
asgiref==3.2.3
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.3
django-decouple==2.1
django-pagedown==2.0.3
django-tinymce==2.8.0
markdown2==2.3.8
pkg-resources==0.0.0
pytz==2019.3
sqlparse==0.3.0
static3==0.7.0
